pyproject.toml
[project]
name = "paylater1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "cryptography>=46.0.3",
    "dotenv>=0.9.9",
    "fastapi[standard]>=0.128.0",
    "passlib[bcrypt]>=1.7.4",
    "pymysql>=1.1.2",
    "python-dotenv>=1.2.1",
    "python-jose>=3.5.0",
    "sqlalchemy>=2.0.45",
    "sqlalchemy-utils>=0.42.1",
]
